completed

+++ CMS_1/admin/includes/view_all_users.php
@@ -8,10 +8,10 @@
             <th>Email</th>
             <th>Role</th>
         </tr>
-    </thead>   
-    
+    </thead>
+
     <tbody>
-        <?php 
+        <?php
 
             $query = "SELECT * FROM users";
             $select_users = mysqli_query($connection, $query);
@@ -30,34 +30,34 @@
                 echo "<td>$user_id </td>";
                 echo "<td>$user_name</td>";
                 echo "<td>$user_firstname</td>";
-                    
-        //        
+
+        //
         //        $query = "SELECT * FROM categories WHERE cat_id = {$post_category_id} ";
-        //        $select_categories_id = mysqli_query($connection,$query);  
+        //        $select_categories_id = mysqli_query($connection,$query);
         //
         //        while($row = mysqli_fetch_assoc($select_categories_id)) {
         //        $cat_id = $row['cat_id'];
         //        $cat_title = $row['cat_title'];
         //
-        //        
+        //
         //        echo "<td>{$cat_title}</td>";
-        //            
+        //
         //        }
-        //        
+        //
                 echo "<td>$user_lastname</td>";
                 echo "<td>$user_email</td>";
                 echo "<td>$user_role</td>";
-                
-                
+
+
         //        $query = "SELECT * FROM posts WHERE post_id = $comment_post_id ";
         //        $select_post_id_query = mysqli_query($connection,$query);
         //        while($row = mysqli_fetch_assoc($select_post_id_query)){
         //        $post_id = $row['post_id'];
         //        $post_title = $row['post_title'];
-        //            
+        //
         //            echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a></td>";
-        //        
-        //        
+        //
+        //
         //        }
 
                 echo "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
@@ -97,19 +97,19 @@
 
     if (isset($_GET['delete'])) {
 
-        //if (isset($_SESSION['user_role'])) {
+        if (isset($_SESSION['user_role'])) { // to aviod sql injection by using delete route directly
 
-            //if ($_SESSION['user_role'] == 'admin') {
+            if ($_SESSION['user_role'] == 'admin') {
 
                 $the_user_id = escape($_GET['delete']);
                 $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
                 $delete_user_query = mysqli_query($connection, $query);
                 header("Location: users.php");
 
-            //}
+            }
 
-        //}
+        }
 
     }
 
-?>    
+?>
